add-docker-compose

+++ Services/eshop.services.CouponAPI/Models/Coupon.cs
@@ -4,7 +4,7 @@
     {
         public int CouponId { get; set; }
         public string? CouponCode { get; set; }
-        public double DiscountAmount { get; set; }
-        public double MinAmount { get; set; }
+        public float DiscountAmount { get; set; }
+        public float MinAmount { get; set; }
     }
 }
