+++ Services/eshop.services.CouponAPI/Models/Dtos/CouponDto.cs
@@ -4,7 +4,7 @@ namespace eshop.services.CouponAPI.Models.Dto
     {
         public int CouponId { get; set; }
         public string? CouponCode { get; set; }
-        public double DiscountAmount { get; set; }
-        public double MinAmount { get; set; }
+        public float DiscountAmount { get; set; }
+        public float MinAmount { get; set; }
     }
 }
